ltc 1.18-rc2

+++ src/ltc/math/radix_to_bin.c
@@ -31,7 +31,7 @@
 
    @return CRYPT_OK on success.
 */
-int radix_to_bin(const void *in, int radix, void *out, size_t* len)
+int radix_to_bin(const void *in, int radix, void *out, unsigned long *len)
 {
    size_t l;
    void* mpi;
